requirements.txt
Django==5.0.4
djangorestframework==3.15.1
channels==4.1.0
websockets==12.0
twilio==9.2.3
python-dotenv==1.0.1
pandas==2.2.2

